pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arc-eval"
version = "0.1.0"
description = "ARC-Eval: Agent Reliability & Compliance evaluation platform for LLMs and AI agents"
readme = "README.md"
license = "MIT"
authors = [{name = "AgentEval Team"}]
keywords = ["ai", "llm", "agent", "evaluation", "compliance", "security", "finance", "ml", "audit"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "reportlab>=4.0.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/arc-computer/arc-eval"
Repository = "https://github.com/arc-computer/arc-eval"
Documentation = "https://github.com/arc-computer/arc-eval/blob/main/README.md"

[project.scripts]
arc-eval = "agent_eval.cli:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true